Added Separate in memory test database for integration tests. Using spring profile "test" and created application-test.properties for test db configuration.
Modified type of field impressions in Campaign class as it does not match type in db. After setting ddl-auto to validate in application.properties

+++ src/main/java/com/optily/campaign_optimisation/entity/Campaign.java
@@ -38,7 +38,7 @@ public class Campaign {
 
     @NonNull
     @Column(name="IMPRESSIONS")
-    Double impressions;
+    Long impressions;
 
     @NonNull
     @Column(name="REVENUE")
